Opslaan functie werkend
Door de methoden binnen de klasse SchetsWin te plaatsen lost de error op
- toegevoegde wijzigingen zijn beschreven in de tekstfile Changes
