+++ Imperatief Programmeren - Schets/SchetsEditor2008.csproj
@@ -54,7 +54,6 @@
     <Compile Include="SchetsControl.cs">
       <SubType>UserControl</SubType>
     </Compile>
-    <Compile Include="SchetsOpslaan.cs" />
     <Compile Include="SchetsWin.cs">
       <SubType>Form</SubType>
     </Compile>
